authorization update/destroy contractor

+++ app/controllers/api/contractors_controller.rb
@@ -13,22 +13,30 @@ class Api::ContractorsController < ApplicationController
     end
 
     def update
-        contractor = Contractor.find_by(id: params[:id])
-        if contractor
-            contractor.update(user_params)
-            render json: contractor
-        else
-            render json: { error: "Profile not found" }, status: :not_found
-        end
+        if current_user.type == "Contractor"
+            contractor = Contractor.find_by(id: params[:id])
+            if contractor
+                 contractor.update(user_params)
+                 render json: contractor
+            else
+                 render json: { error: "Profile not found" }, status: :not_found
+            end
+         else
+             render json: { error: "Unauthorized"}, status: :unauthorized
+         end
     end
 
     def destroy
-        contractor = Contractor.find_by(id: params[:id])
-        if contractor
-            contractor.destroy
-            render json: contractor
+        if current_user.type == "Contractor"
+            contractor = Contractor.find_by(id: params[:id])
+            if contractor
+                contractor.destroy
+                render json: contractor
+            else
+                render json: { error: "Profile not found" }, status: :not_found
+            end
         else
-            render json: { error: "Profile not found" }, status: :not_found
+            render json: { error: "Unauthorized"}, status: :unauthorized
         end
     end
